% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIscreenR-package.R
\docType{package}
\name{AIscreenR-package}
\alias{AIscreenR}
\alias{AIscreenR-package}
\title{AIscreenR: AI Screening Tools in R for Systematic Reviewing}
\description{
Provides functions to conduct title and abstract screening in systematic reviews using large language models, such as the Generative Pre-trained Transformer (GPT) models from 'OpenAI' \url{https://developers.openai.com/}. These functions can enhance the quality of title and abstract screenings while reducing the total screening time significantly. In addition, the package includes tools for quality assessment of title and abstract screenings, as described in Vembye, Christensen, Mølgaard, and Schytt (2025) \doi{10.1037/met0000769}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikkelvembye.github.io/AIscreenR/}
  \item \url{https://github.com/MikkelVembye/AIscreenR}
  \item Report bugs at \url{https://github.com/MikkelVembye/AIscreenR/issues}
}

}
\author{
\strong{Maintainer}: Mikkel H. Vembye \email{mikkel.vembye@gmail.com} (\href{https://orcid.org/0000-0001-9071-0724}{ORCID})

Authors:
\itemize{
  \item Mikkel H. Vembye \email{mikkel.vembye@gmail.com} (\href{https://orcid.org/0000-0001-9071-0724}{ORCID})
  \item Thomas Olsen \email{thomasolsen123@hotmail.com}
}

}
\keyword{internal}
