% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claude_model_prizes}
\alias{claude_model_prizes}
\title{Claude model prices (last updated May 13, 2026)
Data set containing input and output prizes for all Claude's API models.}
\format{
A \code{data.frame} containing 4 rows/models and 3 variables/columns
\tabular{lll}{
\bold{model} \tab \code{character} \tab indicating the specific GPT model \cr
\bold{price_in_per_token} \tab \code{character} \tab indicating the input prize per token \cr
\bold{price_out_per_token} \tab \code{character} \tab indicating the output prize per token \cr
}
}
\usage{
claude_model_prizes
}
\description{
Claude model prices (last updated May 13, 2026)
Data set containing input and output prizes for all Claude's API models.
}
\references{
Anthropic. \emph{Pricing}. \url{https://platform.claude.com/docs/en/about-claude/pricing}
}
\keyword{datasets}
