% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groq_pricing.r
\name{estimate_groq_cost}
\alias{estimate_groq_cost}
\title{Calculate estimated cost for Groq screening}
\usage{
estimate_groq_cost(
  n_studies,
  avg_tokens_per_study = 300,
  avg_completion_tokens = 10,
  model = "llama3-70b-8192",
  reps = 1
)
}
\arguments{
\item{n_studies}{Number of studies to screen}

\item{avg_tokens_per_study}{Average number of tokens per study (input)}

\item{avg_completion_tokens}{Average completion tokens per response}

\item{model}{Groq model name}

\item{reps}{Number of repetitions per study}
}
\value{
Estimated cost in USD
}
\description{
Estimate the cost of screening based on number of studies and tokens
}
\examples{
\dontrun{
estimate_groq_cost(
  n_studies = 100,
  avg_tokens_per_study = 300,
  model = "llama3-70b-8192"
)
}
}
