% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FFT_dat}
\alias{FFT_dat}
\title{Functional Family Therapy}
\format{
A tibble with 150 rows/studies and 4 variables/columns

\tabular{lll}{
\bold{studyid} \tab \code{numeric} \tab Numerical value indicating a unique study-ID
for each study \cr
\bold{title} \tab \code{character} \tab Insert \cr
\bold{abstract}  \tab \code{character}   \tab Insert \cr
\bold{human_code} \tab \code{character}   \tab Insert \cr
}
}
\usage{
FFT_dat
}
\description{
Toy data from a meta-analysis of Functional Familty Therapy
}
\references{
Filges, T., Andersen, D, & Jørgensen, A-M. K (2015).
Functional Family Therapy (FFT) for Young People in Treatment for Non-opioid Drug Use: A Systematic Review
\emph{Campbell Systematic Reviews}, \doi{10.4073/csr.2015.14}
}
\keyword{datasets}
