% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key_functions.R
\name{get_api_key_anthropic}
\alias{get_api_key_anthropic}
\title{Get Anthropic API key from R environment variable.}
\usage{
get_api_key_anthropic(env_var = "ANTHROPIC_API_KEY")
}
\arguments{
\item{env_var}{Character string indicating the name of the temporary R environment variable with
the API key and the used AI model. Currently, the argument only takes \code{env_var = "ANTHROPIC_API_KEY"}.
See \code{\link[=set_api_key]{set_api_key()}} to set/create this variable.}
}
\value{
The specified API key (NOTE: Avoid exposing this in the console).
}
\description{
Get Anthropic API key from R environment variable.
}
\details{
\code{get_api_key_anthropic()} can be used after executing \code{\link[=set_api_key]{set_api_key()}} or by
adding the api key permanently to your R environment by using \code{\link[usethis:edit_r_environ]{usethis::edit_r_environ()}}.
Then write \verb{ANTHROPIC_API_KEY=[insert your api key here]} and close the \code{.Renviron} window and restart R.
}
\note{
Find your personal API key via the Anthropic at \url{https://platform.claude.com/settings/keys}.
}
\examples{

\dontrun{
get_api_key_anthropic()
}
}
\seealso{
\code{\link{set_api_key}}.
}
