% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groq_helpers.r
\name{groq_is_transient}
\alias{groq_is_transient}
\title{Check if error is transient for Groq API}
\usage{
groq_is_transient(resp)
}
\arguments{
\item{resp}{HTTP response object from httr2}
}
\value{
Logical indicating if error is transient and should be retried
}
\description{
Predicate function to determine if an HTTP error should trigger a retry
}
\examples{
\dontrun{
# Used internally in tabscreen_groq for retry logic
groq_is_transient(response)
}
}
