% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groq_model_prizes}
\alias{groq_model_prizes}
\title{Groq model prices (last updated September 1, 2025)}
\format{
A \code{data.frame} containing 4 rows/models and 3 variables/columns

\tabular{lll}{
\bold{model} \tab \code{character} \tab indicating the specific GPT model \cr
\bold{price_in_per_token} \tab \code{character} \tab indicating the input prize per token \cr
\bold{price_out_per_token} \tab \code{character} \tab indicating the output prize per token \cr
}
}
\usage{
groq_model_prizes
}
\description{
Data set containing input and output prizes for all GROQ's API models.
}
\references{
GROQ. \emph{Pricing}. \url{https://groq.com/pricing}
}
\keyword{datasets}
