% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.inherits.R
\name{is_chatgpt_tbl}
\alias{is_chatgpt_tbl}
\title{Test if the object is a \code{'chatgpt_tbl'} object}
\usage{
is_chatgpt_tbl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{chatgpt_tbl} class.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\if{html}{\out{<br>}}
\if{html}{\out{<br>}}
This function returns \code{TRUE} for \code{chatgpt_tbl} objects,
and \code{FALSE} for all other objects.
}
