% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.inherits.R
\name{is_gpt}
\alias{is_gpt}
\title{Test if the object is a \code{'gpt'} object}
\usage{
is_gpt(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{gpt} class.
}
\description{
This function returns \code{TRUE} for \code{gpt} objects,
and \code{FALSE} for all other objects.
}
