% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.inherits.R
\name{is_gpt_tbl}
\alias{is_gpt_tbl}
\title{Test if the object is a \code{'gpt_tbl'} object}
\usage{
is_gpt_tbl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{gpt_tbl} class.
}
\description{
This function returns \code{TRUE} for \code{gpt_tbl} objects,
and \code{FALSE} for all other objects.
}
