% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.inherits.R
\name{is_groq_agg_tbl}
\alias{is_groq_agg_tbl}
\title{Test if the object is a \code{'groq_agg_tbl'} object
This function returns \code{TRUE} for \code{groq_agg_tbl} objects,
and \code{FALSE} for all other objects.}
\usage{
is_groq_agg_tbl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the `groq_agg_tbl
}
\description{
Test if the object is a \code{'groq_agg_tbl'} object
This function returns \code{TRUE} for \code{groq_agg_tbl} objects,
and \code{FALSE} for all other objects.
}
