% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.inherits.R
\name{is_groq_tbl}
\alias{is_groq_tbl}
\title{Test if the object is a \code{'groq'} object}
\usage{
is_groq_tbl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{groq} class.
}
\description{
This function returns \code{TRUE} for \code{groq} objects,
and \code{FALSE} for all other objects.
}
