% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.inherits.R
\name{is_ollama_tbl}
\alias{is_ollama_tbl}
\title{Test if the object is a \code{'ollama'} object}
\usage{
is_ollama_tbl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{ollama} class.
}
\description{
This function returns \code{TRUE} for \code{ollama} objects,
and \code{FALSE} for all other objects.
}
