% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mistral_model_prizes}
\alias{mistral_model_prizes}
\title{Mistral model prices (last updated May 7, 2026)
Data set containing input and output prizes for all Mistral's API models.}
\format{
A \code{data.frame} containing 4 rows/models and 3 variables/columns
\tabular{lll}{
\bold{model} \tab \code{character} \tab indicating the specific GPT model \cr
\bold{price_in_per_token} \tab \code{character} \tab indicating the input prize per token \cr
\bold{price_out_per_token} \tab \code{character} \tab indicating the output prize per token \cr
}
}
\usage{
mistral_model_prizes
}
\description{
Mistral model prices (last updated May 7, 2026)
Data set containing input and output prizes for all Mistral's API models.
}
\references{
Mistral. \emph{Pricing}. \url{https://mistral.ai/pricing}
}
\keyword{datasets}
