% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_prizes}
\alias{model_prizes}
\title{Model prize data (last updated July 31, 2026)}
\format{
A \code{data.frame} containing 42 rows/models and 3 variables/columns

\tabular{lll}{
\bold{model} \tab \code{character} \tab indicating the specific GPT model \cr
\bold{price_in_per_token} \tab \code{character} \tab indicating the input prize per token \cr
\bold{price_out_per_token} \tab \code{character} \tab indicating the output prize per token \cr
}
}
\usage{
model_prizes
}
\description{
Dataset mainly containing input and output prizes for all OpenAI's GPT API models.
}
\references{
OpenAI. \emph{Pricing}. \url{https://developers.openai.com/api/docs/pricing}
}
\keyword{datasets}
