% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gpt.R
\name{print.gpt}
\alias{print.gpt}
\title{Print methods for \code{'gpt'} objects}
\usage{
\method{print}{gpt}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'gpt'}.}

\item{...}{other print arguments.}
}
\value{
Information about how to find answer data sets and pricing information.
}
\description{
Print methods for \code{'gpt'} objects
}
\examples{
\dontrun{
print(x)
}
}
