% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gpt_price.R
\name{print.gpt_price}
\alias{print.gpt_price}
\title{Print methods for \code{'gpt_price'} objects}
\usage{
\method{print}{gpt_price}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"gpt_price"}.}

\item{...}{other print arguments.}
}
\value{
The total price of the screening across all gpt-models expected to be used for the screening.
}
\description{
Print methods for \code{'gpt_price'} objects
}
\examples{
\dontrun{
print(x)
}
}
