% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.groq.r
\name{print.groq}
\alias{print.groq}
\title{Print method for groq objects}
\usage{
\method{print}{groq}(x, ...)
}
\arguments{
\item{x}{A groq object from \code{\link[=tabscreen_groq]{tabscreen_groq()}}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
Invisibly returns the input object \code{x}. Provides the user with information on how to access the results.
}
\description{
Print method for groq objects
}
\examples{
\dontrun{
# res is an object created by tabscreen_groq()
print(res)
}
}
