% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groq_helpers.r
\name{status_code}
\alias{status_code}
\title{Get HTTP status code from last response}
\usage{
status_code()
}
\value{
Integer HTTP status code, or 999 if no response available
}
\description{
Internal function to get the HTTP status code from the most recent API call
}
\keyword{internal}
