% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABscreen_chatgpt.R
\name{tabscreen_chatgpt}
\alias{tabscreen_chatgpt}
\title{Title and abstract screening with ChatGPT}
\usage{
tabscreen_chatgpt(
  data,
  prompt,
  studyid,
  title,
  abstract,
  api_key,
  model = "gpt-3.5-turbo",
  sleep_time = 0,
  time_info = FALSE,
  reps = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{Insert}

\item{prompt}{Insert}

\item{studyid}{Insert}

\item{title}{Insert}

\item{abstract}{Insert}

\item{api_key}{Numerical value with your personal API key. Find at \url{https://platform.openai.com/account/api-keys}}

\item{model}{Character indicating the ChatGPT model to be use. Default is "gpt-3.5-turbo".}

\item{sleep_time}{Numerical value indicating in seconds the sleeping time in between questions. This
is especially helpful when not a pay-as-you-go user. For more information see
\url{https://platform.openai.com/docs/guides/rate-limits/what-are-the-rate-limits-for-our-api}}

\item{time_info}{Logical indicating if the time of the answer should be returned.}

\item{reps}{Numerical indicating the number of times the same question should be sent to ChatGPT.
This can be useful to test consistency between answers. Default is 1.}

\item{seed}{Numerical value for a seed to ensure that proper,
parallel-safe random numbers are produced.}

\item{...}{Further time functions to be added to RETRY.
See \url{https://httr.r-lib.org/reference/RETRY.html}}
}
\value{
A \code{tibble} with answer
}
\description{
This function supports the conduct of title and abstract screening with ChatGPT in R.
The function allow to run title and abstract screening across multiple prompts and with
repeated questions to check for consistency across answers.
}
\examples{

\dontrun{
library(future)

# Find your api key at https://platform.openai.com/account/api-keys
api_key <- 123456789

data <- load("data.RData")

plan(multisession, workers = 7)

system.time(
 test_dat <-
  tabscreen_chatgpt_engine(
    data = data,
    prompt = prompts,
    studyid = studyid,
    title = Title,
    abstract = Abstract,
    api_key = api_key,
    reps = 2
 )
)

}

}
