% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIscreenR-package.R
\docType{package}
\name{AIscreenR-package}
\alias{AIscreenR}
\alias{AIscreenR-package}
\title{AIscreenR: AI Screening Tools in R for Systematic Reviewing}
\description{
A collection of functions to screen scientific studies with AI models. The package includes functions to conduct title and abstract screening with large-language models from 'OpenAI'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikkelvembye.github.io/AIscreenR/}
  \item \url{https://github.com/MikkelVembye/AIscreenR}
  \item Report bugs at \url{https://github.com/MikkelVembye/AIscreenR/issues}
}

}
\author{
\strong{Maintainer}: Mikkel H. Vembye \email{mikkel.vembye@gmail.com} (\href{https://orcid.org/0000-0001-9071-0724}{ORCID}) [copyright holder]

}
\keyword{internal}
