% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIscreenR-package.R
\docType{package}
\name{AIscreenR-package}
\alias{AIscreenR}
\alias{AIscreenR-package}
\title{AIscreenR: AI screening tools for systematic reviews}
\description{
Screening with AI tools, including GPT API models from 'OpenAI'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikkelvembye.github.io/AIscreenR/}
  \item Report bugs at \url{https://github.com/MikkelVembye/AIscreenR/issues}
}

}
\author{
\strong{Maintainer}: Mikkel H. Vembye \email{mikkel.vembye@gmail.com} (\href{https://orcid.org/0000-0001-9071-0724}{ORCID})

}
\keyword{internal}
