% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_price.R
\name{approximate_price_gpt}
\alias{approximate_price_gpt}
\title{Approximate price estimation for title and abstract screening using ChatGPT models}
\usage{
approximate_price_gpt(
  data,
  prompt,
  studyid,
  title,
  abstract,
  model = "gpt-3.5-turbo-0613",
  reps = 1,
  token_word_ratio = 1.6
)
}
\arguments{
\item{data}{Dataset containing the titles and abstracts.}

\item{prompt}{Prompt(s) to paste before the title and abstract.}

\item{studyid}{Unique ID of study. If missing, this is generated
automatically.}

\item{title}{Name of the variable containing the title information.}

\item{abstract}{Name of variable containing the abstract information.}

\item{model}{Character string with the name of the completion model. Can take
multiple models, including gpt-4 models. Default = \code{"gpt-3.5-turbo-0613"}.
Find available model at
\url{https://platform.openai.com/docs/models/model-endpoint-compatibility}.}

\item{reps}{Numerical value indicating the number of times the same
question should be sent to ChatGPT. This can be useful to test consistency
between answers. Default is \code{1} but when using 3.5 models, we recommend setting this
value to \code{10}.}

\item{token_word_ratio}{The multiplier used to approximate the number of tokens per word.
Default is \code{1.6} which we empirically have found to be the average number of tokens per word.}
}
\value{
An object of class \code{"gpt_price"}. The object is a list containing the following
components:
\item{price}{numerical value indicating the total approximate price (in USD) of the screening across all gpt-models expected to be used for the screening.}
\item{price_data}{dataset with prices across all gpt models expected to be used for screening.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\if{html}{\out{<br>}}
\if{html}{\out{<br>}}
This function supports the approximation of the price of the total title and abstract
screening with ChatGPT. The function only provide approximately accurate pricing
estimates for screenings drawing on the function calling function \code{incl_function_simple} (Default)
since the detailed descriptions will increase the completion tokens with an unknown amount
which furthermore is rather model sensitive with gpt-4 models yielding longer
detailed responses relative to gpt-3.5-turbo models.
}
\examples{
prompt <- "This is a prompt"

app_price <- approximate_price_gpt(
  data = FFT_dat[1:2,],
  prompt = prompt,
  studyid = studyid,
  title = title,
  abstract = abstract,
  model = c("gpt-3.5-turbo-0613", "gpt-4"),
  reps = 10
)

app_price
app_price$price_dollar
app_price$price_data
}
