% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABscreen_chatgpt.R
\name{ask_chatgpt}
\alias{ask_chatgpt}
\title{Abstract screening with ChatGPT function}
\usage{
ask_chatgpt(
  question,
  api_key,
  model = "gpt-3.5-turbo",
  sleep_time = 20,
  add_name_to_vars = NULL
)
}
\arguments{
\item{question}{Insert}

\item{api_key}{Insert}

\item{model}{Insert}

\item{sleep_time}{Insert}

\item{add_name_to_vars}{Insert}
}
\value{
A tibble with answer and its running time.
}
\description{
Abstract screening with ChatGPT function
}
\examples{

\dontrun{
q <- "What is a carrot?"
ask_chatgpt(q, api_key = 123456789, sleep_time = 0)
}

}
