% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABscreen_chatgpt.R
\name{ask_chatgpt}
\alias{ask_chatgpt}
\title{Asking a single question to ChatGPT}
\usage{
ask_chatgpt(
  question,
  api_key,
  model = "gpt-3.5-turbo",
  sleep_time = 0,
  time_info = FALSE,
  reps = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{question}{Character with the question you want ChatGPT to answer.}

\item{api_key}{Numerical value with your personal API key. Find at \url{https://platform.openai.com/account/api-keys}}

\item{model}{Character indicating the ChatGPT model to be use. Default is "gpt-3.5-turbo".}

\item{sleep_time}{Numerical value indicating in seconds the sleeping time in between questions. This
is especially helpful when using the function together with \code{purrr::map_*} functions. With a free subscription at ChatGPT
you can ask 3 questions per minute. With a plus subscription you have access to 60 questions per minute, and can reduce the
default sleeping time. Default is 20 seconds.}

\item{time_info}{Logical indicating if the time of the answer should be returned.}

\item{reps}{Numerical indicating the number of times the same questions should be sent to ChatGPT.
This can be useful when investigating the consistency of the yield answer. Default is 1.}

\item{seed}{Numerical value for a seed to ensure that proper,
parallel-safe random numbers are produced.}

\item{...}{Further time functions to be added to the RETRY.
See \url{https://httr.r-lib.org/reference/RETRY.html}}
}
\value{
A tibble with the ChatGPT answer to your question.
}
\description{
Asking a single question to ChatGPT
}
\examples{

\dontrun{

# Find your api key at https://platform.openai.com/account/api-keys
api_key <- 123456789

q <- "What is a carrot?"
ask_chatgpt(q, api_key = api_key, sleep_time = 0)

}

}
