% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabscreen_chatgpt.R
\name{is_chatgpt}
\alias{is_chatgpt}
\title{Test if the object is a chatgpt object}
\usage{
is_chatgpt(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{chatgpt} class.
}
\description{
This function returns \code{TRUE} for \code{chatgpt} objects,
and \code{FALSE} for all other objects.
}
