% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groq_helpers.r
\name{is_groq_tbl}
\alias{is_groq_tbl}
\title{Test if object is groq_tbl}
\usage{
is_groq_tbl(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE if object inherits from groq_tbl class, FALSE otherwise
}
\description{
Check if an object inherits from the groq_tbl class
}
\examples{
\dontrun{
result <- tabscreen_groq(data, prompt, studyid, title, abstract)
is_groq_tbl(result$answer_data)  # TRUE
}
}
