% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.chatgpt.R
\name{print.chatgpt}
\alias{print.chatgpt}
\title{Print methods for 'chatgpt' objects}
\usage{
\method{print}{chatgpt}(x, ...)
}
\arguments{
\item{x}{an object of class "chatgpt".}

\item{...}{other print arguments.}
}
\value{
Information about how to find answer datasets and pricing information.
}
\description{
Print methods for 'chatgpt' objects
}
\examples{
\dontrun{
print(x)
}
}
