% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.groq.r
\name{print.groq}
\alias{print.groq}
\title{Print method for 'groq' objects}
\usage{
\method{print}{groq}(x, ...)
}
\arguments{
\item{x}{A groq object from \code{\link[=tabscreen_groq]{tabscreen_groq()}}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
Information about how to find answer data sets and pricing information.
}
\description{
Print method for 'groq' objects
}
\examples{
\dontrun{
print(x)
}
}
