% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limits.R
\name{rate_limits_per_minute}
\alias{rate_limits_per_minute}
\title{Find updated rate limits for API models}
\usage{
rate_limits_per_minute(
  AI_tool = "CHATGPT",
  model = "gpt-3.5-turbo-16k-0613",
  api_key = get_api_key()
)
}
\arguments{
\item{AI_tool}{Character string specified the AI tool from which the api is
issued. Default is \code{"CHATGPT"}}

\item{model}{Character string with the name of the completion model.
Default = \code{"gpt-3.5-turbo-16k-0613"}.
Find available model at
\url{https://platform.openai.com/docs/models/model-endpoint-compatibility}.}

\item{api_key}{Numerical value with your personal API key. Find at
\url{https://platform.openai.com/account/api-keys}. Use
\code{\link[=secret_make_key]{secret_make_key()}}, \code{\link[=secret_encrypt]{secret_encrypt()}}, and
\code{\link[=secret_decrypt]{secret_decrypt()}} to scramble and decrypt the api key and
use \code{\link[=set_api_key]{set_api_key()}} to securely automate the use of the
api key by setting the api key as a locale environment variable.}
}
\value{
A tibble including variables with information about the model used,
the number of requests and tokens per minute.
}
\description{
\code{rate_limits_per_minute} reports the rate limits for a given api.
includes the allowed requests per minute (RPM) as well as tokens per minute (TPM).
Find general information at
\url{https://platform.openai.com/docs/guides/rate-limits/overview}.
}
\examples{
\dontrun{
set_api_key()

rate_limits_per_minute()
}
}
