% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_errors.gpt.R
\name{screen_errors.gpt}
\alias{screen_errors.gpt}
\title{Re-screen failed requests.}
\usage{
\method{screen_errors}{gpt}(
  object,
  api_key = get_api_key(),
  max_tries = 4,
  max_seconds,
  is_transient,
  backoff,
  after,
  ...
)
}
\arguments{
\item{object}{An object of class \code{'gpt'}.}

\item{api_key}{Numerical value with your personal API key.}

\item{max_tries, max_seconds}{'Cap the maximum number of attempts with
\code{max_tries} or the total elapsed time from the first request with
\code{max_seconds}. If neither option is supplied (the default), \code{\link[=req_perform]{req_perform()}}
will not retry' (Wickham, 2023). Default \code{max_tries} is 4. If missing, the value of \code{max_seconds}
from the original screening (e.g., conducted with \code{\link[=tabscreen_gpt]{tabscreen_gpt()}}) will be used.}

\item{is_transient}{'A predicate function that takes a single argument
(the response) and returns \code{TRUE} or \code{FALSE} specifying whether or not
the response represents a transient error' (Wickham, 2023). If missing, the \code{is_transient}
function from the original screening (e.g., conducted with \code{\link[=tabscreen_gpt]{tabscreen_gpt()}}) will be used.}

\item{backoff}{'A function that takes a single argument (the number of failed
attempts so far) and returns the number of seconds to wait' (Wickham, 2023).
If missing, the \code{backoff}value from the original screening (e.g., conducted with \code{\link[=tabscreen_gpt]{tabscreen_gpt()}}) will be used.}

\item{after}{'A function that takes a single argument (the response) and
returns either a number of seconds to wait or \code{NULL}, which indicates
that a precise wait time is not available that the \code{backoff} strategy
should be used instead' (Wickham, 2023). If missing, the \code{after} value
from the original screening (e.g., conducted with \code{\link[=tabscreen_gpt]{tabscreen_gpt()}}) will be used.}

\item{...}{Further argument to pass to the request body. See \url{https://platform.openai.com/docs/api-reference/chat/create}.
If used in the orginal screening (e.g., with \code{\link[=tabscreen_gpt]{tabscreen_gpt()}}), the argument(s)
must be specified again here.}
}
\value{
Object of class \code{'gpt'} similar to the object returned by \code{\link[=tabscreen_gpt.tools]{tabscreen_gpt.tools()}}.
See documentation for \code{\link[=tabscreen_gpt.tools]{tabscreen_gpt.tools()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\if{html}{\out{<br>}}
\if{html}{\out{<br>}}

This function supports re-screening of all failed title and abstract requests
screened with \code{\link[=tabscreen_gpt.tools]{tabscreen_gpt.tools()}}.
}
\examples{

\dontrun{
prompt <- "Is this study about a Functional Family Therapy (FFT) intervention?"

obj_with_error <-
  tabscreen_gpt(
    data = filges2015_dat[1:10,],
    prompt = prompt,
    studyid = studyid,
    title = title,
    abstract = abstract,
    model = "gpt-4o"
    )

obj_rescreened <-
  obj_with_error |>
  screen_error()

}
}
\references{
Wickham H (2023).
\emph{httr2: Perform HTTP Requests and Process the Responses}.
https://httr2.r-lib.org, https://github.com/r-lib/httr2.
}
\seealso{
\code{\link[=tabscreen_gpt]{tabscreen_gpt()}}, \code{\link[=tabscreen_gpt.tools]{tabscreen_gpt.tools()}}
}
