% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_errors.R
\name{screen_errors}
\alias{screen_errors}
\title{Generic function to re-screen failed title and abstract requests.}
\usage{
screen_errors(
  object,
  api_key = get_api_key(),
  max_tries = 4,
  max_seconds,
  is_transient,
  backoff,
  after,
  ...
)
}
\value{
An object of class \code{'gpt'} or \code{'chatgpt'} similar to the object returned by \code{\link[=tabscreen_gpt]{tabscreen_gpt()}}.
See documentation for \code{\link[=tabscreen_gpt]{tabscreen_gpt()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\if{html}{\out{<br>}}
\if{html}{\out{<br>}}

This is a generic function to re-screen of failed title and abstract requests.
}
\examples{

\dontrun{
prompt <- "Is this study about a Functional Family Therapy (FFT) intervention?"

obj_with_error <-
  tabscreen_gpt(
    data = filges2015_dat[1:10,],
    prompt = prompt,
    studyid = studyid,
    title = title,
    abstract = abstract,
    model = "gpt-4o-mini"
    )

obj_rescreened <-
  obj_with_error |>
  screen_error()

}

}
\seealso{
\code{\link[=screen_errors.gpt]{screen_errors.gpt()}}, \code{\link[=screen_errors.chatgpt]{screen_errors.chatgpt()}}
}
