% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key_functions.R
\name{set_api_key}
\alias{set_api_key}
\title{Creating a temporary R environment API key variable}
\usage{
set_api_key(key, env_var = "CHATGPT_KEY")
}
\arguments{
\item{key}{Character string with an (ideally encrypt) API key. If not provided it
returns a password box in which the true API key can be entered.}

\item{env_var}{Character string indicating the name of the temporary R environment variables
and the used AI model. Currently, the argument only takes \code{env_var = "CHATGPT_KEY"}.}
}
\value{
A temporary environment variable with the name from \code{env_var}.
If \code{key} is missing, it returns a password box in which the true API key can be entered.
}
\description{
Creating a temporary R environment API key variable
}
\examples{
\dontrun{
set_api_key()
}

}
